Making Linux gcc more happy.
git-svn-id: https://svn.vis.uni-stuttgart.de/utilities/vislib/trunk@276 3ec011a9-ccf5-4689-8f2f-5f4aa021d377

+++ test/testmisc.cpp
@@ -473,7 +473,7 @@ void TestSystemMessage(void) {
 
 void TestPerformanceCounter(void) {
     for (int i = 0; i < 100; i++) {
-        ::_tprintf(_T("%lu\n"), vislib::sys::PerformanceCounter::Query());
+        ::_tprintf(_T("%lu\n"), (unsigned long)vislib::sys::PerformanceCounter::Query());
     }
 }
